pycaalp/freedom_matrices_stl_based/requirements.txt
numpy
pyvista
trimesh
python-fcl
meshlib